Refactor: NetworkClient를 object로 관리하는 것이 아니고 의존성 주입을 하도록 변경
