--- core/network/src/main/java/com/alreadyoccupiedseat/network/NetworkClient.kt
@@ -1,35 +1,0 @@
-package com.alreadyoccupiedseat.network
-
-import com.google.gson.GsonBuilder
-import okhttp3.OkHttpClient
-import okhttp3.logging.HttpLoggingInterceptor
-import retrofit2.Retrofit.*
-import retrofit2.converter.gson.GsonConverterFactory
-import java.util.concurrent.TimeUnit
-
-object NetworkClient {
-
-    private val okhttpClient = OkHttpClient.Builder()
-        .readTimeout(10, TimeUnit.SECONDS)
-        .connectTimeout(10, TimeUnit.SECONDS)
-        .writeTimeout(15, TimeUnit.SECONDS)
-        .addInterceptor(
-            HttpLoggingInterceptor().apply {
-                level = HttpLoggingInterceptor.Level.BODY
-            },
-        )
-        .build()
-
-    private val retrofit = Builder()
-        .baseUrl(API.BASE_URL)
-        .client(okhttpClient)
-        .addConverterFactory(
-            GsonConverterFactory.create(
-                GsonBuilder()
-                    .setLenient()
-                    .create(),
-            ),
-        )
-        .build()
-
-}
